+++ integrations/amazon_bedrock/tests/test_converse_generator.py
@@ -1,0 +1,378 @@
+import json
+from unittest.mock import Mock, patch
+
+import pytest
+from botocore.exceptions import ClientError
+
+from haystack_integrations.common.amazon_bedrock.errors import (
+    AmazonBedrockInferenceError,
+)
+from haystack_integrations.components.generators.amazon_bedrock import (
+    MODEL_CAPABILITIES,
+    AmazonBedrockConverseGenerator,
+    ConverseMessage,
+    ToolConfig,
+)
+from haystack_integrations.components.generators.amazon_bedrock.converse.utils import (
+    ConverseRole,
+    ImageBlock,
+    ImageSource,
+    ToolUseBlock,
+)
+
+
+def get_current_weather(location: str, unit: str = "celsius") -> str:
+    """Get the current weather in a given location"""
+    # This is a mock function, replace with actual API call
+    return f"The weather in {location} is 22 degrees {unit}."
+
+
+def get_current_time(timezone: str) -> str:
+    """Get the current time in a given timezone"""
+    # This is a mock function, replace with actual time lookup
+    return f"The current time in {timezone} is 14:30."
+
+
+def test_to_dict(mock_boto3_session):
+    """
+    Test that the to_dict method returns the correct dictionary without aws credentials
+    """
+    tool_config = ToolConfig.from_functions([get_current_weather, get_current_time])
+
+    generator = AmazonBedrockConverseGenerator(
+        model="anthropic.claude-3-5-sonnet-20240620-v1:0",
+        inference_config={
+            "temperature": 0.1,
+            "maxTokens": 256,
+            "topP": 0.1,
+            "stopSequences": ["\\n"],
+        },
+        tool_config=tool_config,
+    )
+
+    expected_dict = {
+        "type": "haystack_integrations.components.generators.amazon_bedrock.converse.converse_generator."
+        "AmazonBedrockConverseGenerator",
+        "init_parameters": {
+            "aws_access_key_id": {
+                "type": "env_var",
+                "env_vars": ["AWS_ACCESS_KEY_ID"],
+                "strict": False,
+            },
+            "aws_secret_access_key": {
+                "type": "env_var",
+                "env_vars": ["AWS_SECRET_ACCESS_KEY"],
+                "strict": False,
+            },
+            "aws_session_token": {
+                "type": "env_var",
+                "env_vars": ["AWS_SESSION_TOKEN"],
+                "strict": False,
+            },
+            "aws_region_name": {
+                "type": "env_var",
+                "env_vars": ["AWS_DEFAULT_REGION"],
+                "strict": False,
+            },
+            "aws_profile_name": {
+                "type": "env_var",
+                "env_vars": ["AWS_PROFILE"],
+                "strict": False,
+            },
+            "model": "anthropic.claude-3-5-sonnet-20240620-v1:0",
+            "inference_config": {
+                "temperature": 0.1,
+                "maxTokens": 256,
+                "topP": 0.1,
+                "stopSequences": ["\\n"],
+            },
+            "tool_config": tool_config.to_dict(),
+            "streaming_callback": None,
+            "system_prompt": None,
+        },
+    }
+
+    assert generator.to_dict() == expected_dict
+
+
+def test_from_dict(mock_boto3_session):
+    """
+    Test that the from_dict method returns the correct object
+    """
+    tool_config = ToolConfig.from_functions([get_current_weather, get_current_time])
+
+    generator = AmazonBedrockConverseGenerator.from_dict(
+        {
+            "type": "haystack_integrations.components.generators.amazon_bedrock.converse.converse_generator."
+            "AmazonBedrockConverseGenerator",
+            "init_parameters": {
+                "aws_access_key_id": {"type": "env_var", "env_vars": ["AWS_ACCESS_KEY_ID"], "strict": False},
+                "aws_secret_access_key": {"type": "env_var", "env_vars": ["AWS_SECRET_ACCESS_KEY"], "strict": False},
+                "aws_session_token": {"type": "env_var", "env_vars": ["AWS_SESSION_TOKEN"], "strict": False},
+                "aws_region_name": {"type": "env_var", "env_vars": ["AWS_DEFAULT_REGION"], "strict": False},
+                "aws_profile_name": {"type": "env_var", "env_vars": ["AWS_PROFILE"], "strict": False},
+                "model": "anthropic.claude-3-5-sonnet-20240620-v1:0",
+                "inference_config": {
+                    "temperature": 0.1,
+                    "maxTokens": 256,
+                    "topP": 0.1,
+                    "stopSequences": ["\\n"],
+                },
+                "tool_config": tool_config.to_dict(),
+                "streaming_callback": None,
+                "system_prompt": None,
+            },
+        }
+    )
+
+    assert generator.inference_config["temperature"] == 0.1
+    assert generator.inference_config["maxTokens"] == 256
+    assert generator.inference_config["topP"] == 0.1
+    assert generator.inference_config["stopSequences"] == ["\\n"]
+    assert generator.tool_config.to_dict() == tool_config.to_dict()
+    assert generator.model == "anthropic.claude-3-5-sonnet-20240620-v1:0"
+
+
+def test_default_constructor(mock_boto3_session, set_env_variables):
+    """
+    Test that the default constructor sets the correct values
+    """
+
+    layer = AmazonBedrockConverseGenerator(
+        model="anthropic.claude-3-5-sonnet-20240620-v1:0",
+    )
+
+    assert layer.model == "anthropic.claude-3-5-sonnet-20240620-v1:0"
+    assert layer.inference_config is None
+    assert layer.tool_config is None
+    assert layer.streaming_callback is None
+    assert layer.system_prompt is None
+
+    # assert mocked boto3 client called exactly once
+    mock_boto3_session.assert_called_once()
+
+    # assert mocked boto3 client was called with the correct parameters
+    mock_boto3_session.assert_called_with(
+        aws_access_key_id="some_fake_id",
+        aws_secret_access_key="some_fake_key",
+        aws_session_token="some_fake_token",
+        profile_name="some_fake_profile",
+        region_name="fake_region",
+    )
+
+
+def test_constructor_with_empty_model():
+    """
+    Test that the constructor raises an error when the model is empty
+    """
+    with pytest.raises(ValueError, match="cannot be None or empty string"):
+        AmazonBedrockConverseGenerator(model="")
+
+
+def test_get_model_capabilities(mock_boto3_session):
+    generator = AmazonBedrockConverseGenerator(model="anthropic.claude-3-sonnet-20240229-v1:0")
+    assert generator.model_capabilities == MODEL_CAPABILITIES["anthropic.claude-3.*"]
+
+    generator = AmazonBedrockConverseGenerator(model="ai21.j2-ultra-instruct-v1")
+    assert generator.model_capabilities == MODEL_CAPABILITIES["ai21.j2-.*"]
+
+    with pytest.raises(ValueError, match="Unsupported model"):
+        AmazonBedrockConverseGenerator(model="unsupported.model-v1")
+
+
+@patch("boto3.Session")
+def test_run_with_different_message_types(mock_session):
+    mock_client = Mock()
+    mock_session.return_value.client.return_value = mock_client
+    mock_client.converse.return_value = {
+        "output": {"message": {"role": "assistant", "content": [{"text": "Hello, how can I help you?"}]}},
+        "usage": {"inputTokens": 10, "outputTokens": 20},
+        "metrics": {"timeToFirstToken": 0.5},
+    }
+
+    generator = AmazonBedrockConverseGenerator(model="anthropic.claude-3-sonnet-20240229-v1:0")
+
+    messages = [
+        ConverseMessage.from_user(["What's the weather like?"]),
+        ConverseMessage.from_user([ImageBlock(format="png", source=ImageSource(bytes=b"fake_image_data"))]),
+    ]
+
+    result = generator.run(messages)
+
+    assert result["message"].role == ConverseRole.ASSISTANT
+    assert result["message"].content.content[0] == "Hello, how can I help you?"
+    assert result["usage"] == {"inputTokens": 10, "outputTokens": 20}
+    assert result["metrics"] == {"timeToFirstToken": 0.5}
+
+    # Check the actual content sent to the API
+    mock_client.converse.assert_called_once()
+    call_args = mock_client.converse.call_args[1]
+    assert len(call_args["messages"]) == 2
+    assert call_args["messages"][0]["content"] == [{"text": "What's the weather like?"}]
+
+
+def test_streaming(mock_boto3_session):
+    generator = AmazonBedrockConverseGenerator(model="anthropic.claude-3-sonnet-20240229-v1:0")
+
+    mocked_events = [
+        {"messageStart": {"role": "assistant"}},
+        {"contentBlockDelta": {"delta": {"text": "To"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " answer"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " your questions"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": ","}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " I'll"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " need to"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " use"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " two"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " different functions"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": ":"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " one"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " to check"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " the weather"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " in Paris and another"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " to get the current"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " time in New York"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": "."}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " Let"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " me"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " fetch"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " that"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " information for"}, "contentBlockIndex": 0}},
+        {"contentBlockDelta": {"delta": {"text": " you."}, "contentBlockIndex": 0}},
+        {"contentBlockStop": {"contentBlockIndex": 0}},
+        {
+            "contentBlockStart": {
+                "start": {"toolUse": {"toolUseId": "tooluse_5Uu9EPSjQxiSsmc5Ex5MJg", "name": "get_current_weather"}},
+                "contentBlockIndex": 1,
+            }
+        },
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": ""}}, "contentBlockIndex": 1}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": '{"loc'}}, "contentBlockIndex": 1}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": 'ation":'}}, "contentBlockIndex": 1}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": ' "Paris"'}}, "contentBlockIndex": 1}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": ', "u'}}, "contentBlockIndex": 1}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": 'nit": "ce'}}, "contentBlockIndex": 1}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": "lsius"}}, "contentBlockIndex": 1}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": '"}'}}, "contentBlockIndex": 1}},
+        {"contentBlockStop": {"contentBlockIndex": 1}},
+        {
+            "contentBlockStart": {
+                "start": {"toolUse": {"toolUseId": "tooluse_cbK-e15KTFqZHtwpBJ0kzg", "name": "get_current_time"}},
+                "contentBlockIndex": 2,
+            }
+        },
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": ""}}, "contentBlockIndex": 2}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": '{"timezon'}}, "contentBlockIndex": 2}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": 'e"'}}, "contentBlockIndex": 2}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": ': "A'}}, "contentBlockIndex": 2}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": "meric"}}, "contentBlockIndex": 2}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": "a/New"}}, "contentBlockIndex": 2}},
+        {"contentBlockDelta": {"delta": {"toolUse": {"input": '_York"}'}}, "contentBlockIndex": 2}},
+        {"contentBlockStop": {"contentBlockIndex": 2}},
+        {"messageStop": {"stopReason": "tool_use"}},
+        {
+            "metadata": {
+                "usage": {"inputTokens": 446, "outputTokens": 118, "totalTokens": 564},
+                "metrics": {"latencyMs": 3930},
+            }
+        },
+    ]
+
+    mock_stream = Mock()
+    mock_stream.__iter__ = Mock(return_value=iter(mocked_events))
+
+    generator.client.converse_stream = Mock(return_value={"stream": mock_stream})
+
+    chunks = []
+    result = generator.run(
+        [ConverseMessage.from_user(["What's the weather like in Paris and what time is it in New York?"])],
+        streaming_callback=lambda chunk: chunks.append(chunk),
+    )
+
+    assert len(chunks) == len(mocked_events)
+    assert (
+        result["message"].content.content[0]
+        == "To answer your questions, I'll need to use two different functions: one to check the weather "
+        "in Paris and another to get the current time in New York. Let me fetch that information for you."
+    )
+    assert len(result["message"].content.content) == 3
+    assert result["stop_reason"] == "tool_use"
+
+    assert result["message"].role == ConverseRole.ASSISTANT
+
+    assert result["usage"] == {"inputTokens": 446, "outputTokens": 118, "totalTokens": 564}
+    assert result["metrics"] == {"latencyMs": 3930}
+
+    assert result["message"].content.content[1].name == "get_current_weather"
+    assert result["message"].content.content[2].name == "get_current_time"
+
+    assert json.dumps(result["message"].content.content[1].input) == """{"location": "Paris", "unit": "celsius"}"""
+    assert json.dumps(result["message"].content.content[2].input) == """{"timezone": "America/New_York"}"""
+
+
+def test_client_error_handling(mock_boto3_session):
+    generator = AmazonBedrockConverseGenerator(model="anthropic.claude-3-sonnet-20240229-v1:0")
+    generator.client.converse = Mock(
+        side_effect=ClientError(
+            error_response={"Error": {"Code": "ThrottlingException", "Message": "Rate exceeded"}},
+            operation_name="Converse",
+        )
+    )
+
+    with pytest.raises(AmazonBedrockInferenceError, match="Could not run inference on Amazon Bedrock model"):
+        generator.run([ConverseMessage.from_user(["Hi"])])
+
+
+def test_tool_usage(mock_boto3_session):
+    tool_config = ToolConfig.from_functions([get_current_weather, get_current_time])
+    generator = AmazonBedrockConverseGenerator(model="anthropic.claude-3-sonnet-20240229-v1:0", tool_config=tool_config)
+
+    mock_response = {
+        "output": {
+            "message": {
+                "role": "assistant",
+                "content": [
+                    {
+                        "text": "I'll get the weather in Paris and the current time in New York for you. "
+                        "To do this, I'll need to use two different tools. Let me fetch that data."
+                    },
+                    {
+                        "toolUse": {
+                            "toolUseId": "tooluse_-Tp78_OeSq-1DSsP0B__TA",
+                            "name": "get_current_weather",
+                            "input": {"location": "Paris", "unit": "celsius"},
+                        }
+                    },
+                    {
+                        "toolUse": {
+                            "toolUseId": "tooluse_gdYvqeiGTme7toWoV4sSKw",
+                            "name": "get_current_time",
+                            "input": {"timezone": "America/New_York"},
+                        },
+                    },
+                ],
+            },
+        },
+        "stopReason": "tool_use",
+    }
+    generator.client.converse = Mock(return_value=mock_response)
+
+    result = generator.run([ConverseMessage.from_user(["What's the weather in London?"])])
+
+    assert len(result["message"].content.content) == 3
+    assert isinstance(result["message"].content.content[0], str)
+    assert isinstance(result["message"].content.content[1], ToolUseBlock)
+    assert isinstance(result["message"].content.content[2], ToolUseBlock)
+    assert result["stop_reason"] == "tool_use"
+    assert result["message"].role == ConverseRole.ASSISTANT
+    assert result["message"].content.content[0] == (
+        "I'll get the weather in Paris and the current time in New York for you. "
+        "To do this, I'll need to use two different tools. Let me fetch that data."
+    )
+    assert result["message"].content.content[1].name == "get_current_weather"
+    assert result["message"].content.content[2].name == "get_current_time"
+    assert json.dumps(result["message"].content.content[1].input) == """{"location": "Paris", "unit": "celsius"}"""
+    assert json.dumps(result["message"].content.content[2].input) == """{"timezone": "America/New_York"}"""
+    assert result["message"].content.content[1].input["location"] == "Paris"
+    assert result["message"].content.content[1].input["unit"] == "celsius"
+    assert result["message"].content.content[2].input["timezone"] == "America/New_York"